+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ---
 
+#### 23.07.2018
+
+* [Rust] Code of Kutulu - Fixed running away (CodinGame)
+
 #### 22.07.2018
 
 * [Rust] Code of Kutulu - Move dist const (CodinGame)